Update test paths and simplify names
