+++ tests/test_dem_generation_local_1/instruction.json
@@ -9,7 +9,7 @@
 		}
 	},
 	"data_paths": {
-		"local_cache": "tests/test_processor_local_files/data",
+		"local_cache": "tests/test_dem_generation_local_1/data",
 		"subfolder": "results",
 		"catchment_boundary": "catchment_boundary.geojson",
 		"land": "land.geojson",
