+++ tests/test_dem_generation_local_2/test.py
@@ -10,25 +10,30 @@ import json
 import pathlib
 import shutil
 import rioxarray
-import xarray
 import numpy
 import shapely
 import geopandas
 import pdal
+import pytest
+import sys
 import logging
 import gc
 
 from src.geofabrics import processor
 
 
-class ProcessorLocalFilesTest(unittest.TestCase):
-    """Tests the basic DemGenerator processor class for a simple example with land,
-    offshore, a coarse DEM and
-    LiDAR using the data specified in the instruction.json
+class Test(unittest.TestCase):
+    """A class to test the basic DemGenerator processor class for a simple example with
+    land, offshore bathymetry, and LiDAR using the data generated in the set-up routine
+    and referenced in the instruction.json.
+
+    The dem.DenseDem.CACHE_SIZE is exceded and the offshore sampled points are
+    re-sampled at a lower resolution.
 
     Tests run include:
-        1. test_result_dem  Check the generated DEM matches the benchmark DEM within a
-        tolerance
+        1. test_result_dem  Check the generated DEM matches the benchmark DEM exactly in
+           Windows, or approximately in Linux. It aims to ensure the offshore resolution
+           is reduced.
     """
 
     LAS_GROUND = 2
@@ -36,11 +41,10 @@ class ProcessorLocalFilesTest(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         """Create a cache directory and CatchmentGeometry object for use in the tests
-        and also download the files used
-        in the tests."""
+        and also download the files used in the tests."""
 
         test_path = pathlib.Path().cwd() / pathlib.Path(
-            "tests/test_processor_local_files"
+            "tests/test_dem_generation_local_2"
         )
 
         # Setup logging
@@ -50,7 +54,7 @@ class ProcessorLocalFilesTest(unittest.TestCase):
             level=logging.INFO,
             force=True,
         )
-        logging.info("In test_processor_local_files.py")
+        logging.info("In test_processor_local_files_limited_offshore.py")
 
         # Load in the test instructions
         instruction_file_path = test_path / "instruction.json"
@@ -59,101 +63,65 @@ class ProcessorLocalFilesTest(unittest.TestCase):
         # Remove any files from last test, then create a results directory
         cls.cache_dir = test_path / "data"
         cls.results_dir = cls.cache_dir / "results"
-        cls.clean_data_folder()
+        cls.tearDownClass()
         cls.results_dir.mkdir()
 
         # Generate catchment data
         catchment_file = cls.results_dir / "catchment_boundary.geojson"
-        x0 = 250
-        y0 = -250
-        x1 = 1250
-        y1 = 750
+        x0 = 0
+        x1 = 3000
+        y0 = -5
+        y1 = 1
         catchment = shapely.geometry.Polygon([(x0, y0), (x1, y0), (x1, y1), (x0, y1)])
         catchment = geopandas.GeoSeries([catchment])
         catchment = catchment.set_crs(cls.instructions["output"]["crs"]["horizontal"])
         catchment.to_file(catchment_file)
 
-        # Generate land data
+        # Generate land data - wider than catchment
         land_file = cls.results_dir / "land.geojson"
-        x0 = 0
+        x0 = -50
+        x1 = 3050
         y0 = 0
-        x1 = 1500
-        y1 = 1000
+        y1 = 5
         land = shapely.geometry.Polygon([(x0, y0), (x1, y0), (x1, y1), (x0, y1)])
         land = geopandas.GeoSeries([land])
         land = land.set_crs(cls.instructions["output"]["crs"]["horizontal"])
         land.to_file(land_file)
 
-        # Generate bathymetry data
+        # Generate bathymetry data - wider than catchment
         bathymetry_file = cls.results_dir / "bathymetry.geojson"
-        x0 = 0
-        x1 = 1500
-        y0 = -50
-        y1 = -100
-        y2 = -200
-        y3 = -300
+        x0 = -50
+        x1 = 3050
+        y0 = -1
+        y1 = -3
+        y2 = -5
         contour_0 = shapely.geometry.LineString(
-            [(x0, y0, -y0 / 10), (x1, y0, -y0 / 10)]
+            [(x0, y0, -y0 * 0.1), ((x0 + x1) / 2, y0, 0), (x1, y0, -y0 * 0.1)]
         )
         contour_1 = shapely.geometry.LineString(
-            [(x0, y1, -y1 / 10), (x1, y1, -y1 / 10)]
+            [(x0, y1, -y1 * 0.1), ((x0 + x1) / 2, y1, 0), (x1, y1, -y1 * 0.1)]
         )
         contour_2 = shapely.geometry.LineString(
-            [(x0, y2, -y2 / 10), (x1, y2, -y2 / 10)]
+            [(x0, y2, -y2 * 0.1), ((x0 + x1) / 2, y2, 0), (x1, y2, -y2 * 0.1)]
         )
-        contour_3 = shapely.geometry.LineString(
-            [(x0, y3, -y3 / 10), (x1, y3, -y3 / 10)]
-        )
-        contours = geopandas.GeoSeries([contour_0, contour_1, contour_2, contour_3])
+        contours = geopandas.GeoSeries([contour_0, contour_1, contour_2])
         contours = contours.set_crs(cls.instructions["output"]["crs"]["horizontal"])
         contours.to_file(bathymetry_file)
 
-        # Create a coarse DEM
-        dem_file = cls.results_dir / "coarse_dem.nc"
-        dxy = 15
-        grid_dem_x, grid_dem_y = numpy.meshgrid(
-            numpy.arange(200, 1300, dxy), numpy.arange(-25, 800, dxy)
-        )
-        grid_dem_z = numpy.zeros_like(grid_dem_x, dtype=numpy.float64)
-        grid_dem_z[grid_dem_y < 0] = grid_dem_y[grid_dem_y < 0] / 10
-        grid_dem_z[grid_dem_y > 0] = (
-            grid_dem_y[grid_dem_y > 0]
-            / 10
-            * (numpy.abs(grid_dem_x[grid_dem_y > 0] - 750) / 500 + 0.1)
-            / 1.1
-        )
-        dem = xarray.DataArray(
-            grid_dem_z,
-            coords={"x": grid_dem_x[0], "y": grid_dem_y[:, 0]},
-            dims=["y", "x"],
-            attrs={"scale_factor": 1.0, "add_offset": 0.0},
-        )
-        dem.rio.write_crs(
-            cls.instructions["output"]["crs"]["horizontal"],
-            inplace=True,
-        )
-        dem.rio.write_transform(inplace=True)
-        dem.rio.write_nodata(numpy.nan, encoded=True, inplace=True)
-        dem.name = "z"
-        dem.to_netcdf(dem_file, format="NETCDF4", engine="netcdf4")
-        # explicitly free memory as xarray seems to be hanging onto memory
-        dem.close()
-        del dem
-
-        # Create LiDAR
+        # Create LiDAR - wider than catchment
         lidar_file = cls.results_dir / "lidar.laz"
-        dxy = 1
+        x0 = -50
+        x1 = 3050
+        y0 = (
+            -5
+        )  # drop offshore LiDAR is True in the instruction file so only LiDAR in the
+        # foreshore will be kept
+        y1 = 5
+        dxy = 0.1
         grid_lidar_x, grid_lidar_y = numpy.meshgrid(
-            numpy.arange(500, 1000, dxy), numpy.arange(-25, 475, dxy)
+            numpy.arange(x0, x1, dxy), numpy.arange(y0, y1, dxy)
         )
-        grid_lidar_z = numpy.zeros_like(grid_lidar_x, dtype=numpy.float64)
-        grid_lidar_z[grid_lidar_y < 0] = grid_lidar_y[grid_lidar_y < 0] / 10
-        grid_lidar_z[grid_lidar_y > 0] = (
-            grid_lidar_y[grid_lidar_y > 0]
-            / 10
-            * (numpy.abs(grid_lidar_x[grid_lidar_y > 0] - 750) / 500 + 0.1)
-            / 1.1
-        )
+        grid_lidar_z = grid_lidar_y / 10
         lidar_array = numpy.empty(
             [len(grid_lidar_x.flatten())],
             dtype=[("X", "<f8"), ("Y", "<f8"), ("Z", "<f8"), ("Classification", "u1")],
@@ -166,9 +134,8 @@ class ProcessorLocalFilesTest(unittest.TestCase):
         pdal_pipeline_instructions = [
             {
                 "type": "writers.las",
-                "a_srs": f"EPSG:"
-                f"{cls.instructions['output']['crs']['horizontal']}+"
-                f"{cls.instructions['output']['crs']['vertical']}",
+                "a_srs": f"EPSG:{cls.instructions['output']['crs']['horizontal']}+"
+                + f"{cls.instructions['output']['crs']['vertical']}",
                 "filename": str(lidar_file),
                 "compression": "laszip",
             }
@@ -179,7 +146,7 @@ class ProcessorLocalFilesTest(unittest.TestCase):
         )
         pdal_pipeline.execute()
 
-        # Run pipeline
+        # Run geofabrics processing pipeline
         runner = processor.RawLidarDemGenerator(cls.instructions)
         runner.run()
         runner = processor.HydrologicDemGenerator(cls.instructions)
@@ -211,31 +178,63 @@ class ProcessorLocalFilesTest(unittest.TestCase):
                         shutil.rmtree(file)
                 shutil.rmtree(path)
 
-    def test_result_dem(self):
+    @pytest.mark.skipif(sys.platform != "win32", reason="Windows test - this is strict")
+    def test_result_dem_windows(self):
         """A basic comparison between the generated and benchmark DEM"""
 
         # Load in benchmark DEM
         file_path = self.cache_dir / self.instructions["data_paths"]["benchmark_dem"]
         with rioxarray.rioxarray.open_rasterio(file_path, masked=True) as benchmark_dem:
-            benchmark_dem = benchmark_dem.squeeze("band", drop=True)
+            benchmark_dem.load()
         # Load in result DEM
         file_path = self.results_dir / self.instructions["data_paths"]["result_dem"]
         with rioxarray.rioxarray.open_rasterio(file_path, masked=True) as test_dem:
-            test_dem = test_dem.squeeze("band", drop=True)
-        # Compare DEMs z - load both from file as rioxarray.rioxarray.open_rasterio
-        # ignores index order
+            test_dem.load()
+        # Compare DEMs - load from file as rioxarray.rioxarray.open_rasterio ignores
+        # index order
         diff_array = test_dem.z.data - benchmark_dem.z.data
-        logging.info(f"DEM z array diff is: {diff_array[diff_array != 0]}")
+        logging.info(f"DEM array diff is: {diff_array[diff_array != 0]}")
+        numpy.testing.assert_array_equal(
+            test_dem.z.data,
+            benchmark_dem.z.data,
+            err_msg="The generated result_dem differs from the benchmark_dem",
+        )
+
+        # explicitly free memory as xarray seems to be hanging onto memory
+        del test_dem
+        del benchmark_dem
+        gc.collect()
+
+    @pytest.mark.skipif(
+        sys.platform != "linux", reason="Linux test - this is less strict"
+    )
+    def test_result_dem_linux(self):
+        """A basic comparison between the generated and benchmark DEM"""
+
+        # Load in benchmark DEM
+        file_path = self.cache_dir / self.instructions["data_paths"]["benchmark_dem"]
+        with rioxarray.rioxarray.open_rasterio(file_path, masked=True) as benchmark_dem:
+            benchmark_dem.load()
+        # Load in result DEM
+        file_path = self.results_dir / self.instructions["data_paths"]["result_dem"]
+        with rioxarray.rioxarray.open_rasterio(file_path, masked=True) as test_dem:
+            test_dem.load()
+        # Compare DEMs z elevations
+        diff_array = test_dem.z.data - benchmark_dem.z.data
+        logging.info(f"DEM z diff is: {diff_array[diff_array != 0]}")
         numpy.testing.assert_array_almost_equal(
             test_dem.z.data,
             benchmark_dem.z.data,
-            err_msg="The generated result_dem z has different data from the "
-            "benchmark_dem",
+            decimal=4,
+            err_msg="The generated result_dem has different data from the "
+            + "benchmark_dem",
         )
 
         # Compare DEMs source classification
         diff_array = test_dem.source_class.data - benchmark_dem.source_class.data
-        logging.info(f"DEM z array diff is: {diff_array[diff_array != 0]}")
+        logging.info(
+            f"DEM source classification diff is: {diff_array[diff_array != 0]}"
+        )
         numpy.testing.assert_array_almost_equal(
             test_dem.source_class.data,
             benchmark_dem.source_class.data,
@@ -244,8 +243,6 @@ class ProcessorLocalFilesTest(unittest.TestCase):
         )
 
         # explicitly free memory as xarray seems to be hanging onto memory
-        test_dem.close()
-        benchmark_dem.close()
         del test_dem
         del benchmark_dem
         gc.collect()
