+++ tests/test_dem_generation_local_3/instruction.json
@@ -9,7 +9,7 @@
 		}
 	},
 	"data_paths": {
-		"local_cache": "tests/test_processor_local_files_offshore_after/data",
+		"local_cache": "tests/test_dem_generation_local_3/data",
 		"subfolder": "results",
 		"catchment_boundary": "catchment_boundary.geojson",
 		"land": "land.geojson",
