+++ tests/test_dem_generation_wellington_1/test.py
@@ -21,22 +21,32 @@ import gc
 from src.geofabrics import processor
 
 
-class MultipleLidarGroundOnlyTest(unittest.TestCase):
+class Test(unittest.TestCase):
     """A class to test the basic processor class Processor functionality for remote
-    tiles by downloading files from
-    OpenTopography within a small region and then generating a DEM. All files are
-    deleted after checking the DEM.
-
-    Note this spans land and sea in comparison to the companion
-    `test_processor_remote_tiles_wellington` that contains
-    only ocean data. The benchmark shoes buildings are removed. No offshore information
-    is provided.
+    tiles by downloading files from OpenTopography within a small region and then
+    generating a DEM. All files are deleted after checking the DEM.
 
     Tests run include:
-        1. test_result_dem_windows/linux - Check the generated DEM matches the benchmark
+        1. test_correct_dataset - Test that the expected dataset is downloaded from
+           OpenTopography
+        2. test_correct_lidar_files_downloaded - Test the downloaded LIDAR files have
+           the expected names
+        3. test_correct_lidar_file_size - Test the downloaded LIDAR files have the
+           expected file sizes
+        4. test_result_dem_windows/linux - Check the generated DEM matches the benchmark
            DEM, where the rigor of the test depends on the operating system (windows or
                                                                              Linux)
     """
+
+    # The expected datasets and files to be downloaded - used for comparison in tests
+    DATASET = "Wellington_2013"
+    FILE_SIZES = {
+        "ot_CL1_WLG_2013_1km_085033.laz": 6795072,
+        "ot_CL1_WLG_2013_1km_086033.laz": 5712485,
+        "ot_CL1_WLG_2013_1km_085032.laz": 1670549,
+        "ot_CL1_WLG_2013_1km_086032.laz": 72787,
+        DATASET + "_TileIndex.zip": 598532,
+    }
 
     @classmethod
     def setUpClass(cls):
@@ -44,7 +54,7 @@ class MultipleLidarGroundOnlyTest(unittest.TestCase):
         chain to download remote files and produce a DEM prior to testing."""
 
         test_path = pathlib.Path().cwd() / pathlib.Path(
-            "tests/test_multiple_lidar_ground_only"
+            "tests/test_dem_generation_wellington_1"
         )
 
         # Setup logging
@@ -54,7 +64,7 @@ class MultipleLidarGroundOnlyTest(unittest.TestCase):
             level=logging.INFO,
             force=True,
         )
-        logging.info("In test_multiple_lidar_ground_only.py")
+        logging.info("In test_dem_generation_wellington_1")
 
         # load in the test instructions
         instruction_file_path = test_path / "instruction.json"
@@ -67,10 +77,10 @@ class MultipleLidarGroundOnlyTest(unittest.TestCase):
         cls.results_dir.mkdir()
 
         # create fake catchment boundary
-        x0 = 1778250
-        y0 = 5470800
-        x1 = 1778550
-        y1 = 5470500
+        x0 = 1764864
+        y0 = 5470382
+        x1 = 1765656
+        y1 = 5471304
         catchment = shapely.geometry.Polygon([(x0, y0), (x1, y0), (x1, y1), (x0, y1)])
         catchment = geopandas.GeoSeries([catchment])
         catchment = catchment.set_crs(cls.instructions["output"]["crs"]["horizontal"])
@@ -110,6 +120,76 @@ class MultipleLidarGroundOnlyTest(unittest.TestCase):
                         shutil.rmtree(file)
                 shutil.rmtree(path)
 
+    def test_correct_dataset(self):
+        """A test to see if the correct dataset is downloaded"""
+
+        dataset_dir = self.cache_dir / self.DATASET
+
+        # check the right dataset is downloaded - self.DATASET
+        self.assertEqual(
+            len(list(self.cache_dir.glob("*/**"))),
+            2,
+            f"There should only be one dataset named {self.DATASET} as well as the "
+            f" generated results folder {self.results_dir} instead there are "
+            f"{len(list(self.cache_dir.glob('*/**')))} list "
+            f"{list(self.cache_dir.glob('*/**'))}",
+        )
+
+        self.assertEqual(
+            len(
+                [
+                    file
+                    for file in self.cache_dir.iterdir()
+                    if file.is_dir() and file == dataset_dir
+                ]
+            ),
+            1,
+            f"Only the {self.DATASET} directory should have been downloaded. Instead we"
+            f" have: {[file for file in self.cache_dir.iterdir() if file.is_dir()]}",
+        )
+
+    def test_correct_files_downloaded(self):
+        """A test to see if all expected dataset files are downloaded"""
+
+        dataset_dir = self.cache_dir / self.DATASET
+        downloaded_files = [dataset_dir / file for file in self.FILE_SIZES.keys()]
+
+        # check files are correct
+        self.assertEqual(
+            len(list(dataset_dir.glob("*"))),
+            len(downloaded_files),
+            "There should have been "
+            f"{len(downloaded_files)} files downloaded into the {self.DATASET} "
+            f"directory, instead there are {len(list(dataset_dir.glob('*')))} "
+            "files/dirs in the directory",
+        )
+
+        self.assertTrue(
+            numpy.all([file in downloaded_files for file in dataset_dir.glob("*")]),
+            f"The downloaded files {list(dataset_dir.glob('*'))} do not match the "
+            f"expected files {downloaded_files}",
+        )
+
+    def test_correct_file_size(self):
+        """A test to see if all expected dataset files are of the right size"""
+
+        dataset_dir = self.cache_dir / self.DATASET
+        downloaded_files = [dataset_dir / file for file in self.FILE_SIZES.keys()]
+
+        # check sizes are correct
+        self.assertTrue(
+            numpy.all(
+                [
+                    downloaded_file.stat().st_size
+                    == self.FILE_SIZES[downloaded_file.name]
+                    for downloaded_file in downloaded_files
+                ]
+            ),
+            "There is a miss-match between the size of the downloaded files "
+            + f"{[file.stat().st_size for file in downloaded_files]} and the expected "
+            + f"sizes of {self.FILE_SIZES.values()}",
+        )
+
     @pytest.mark.skipif(sys.platform != "win32", reason="Windows test - this is strict")
     def test_result_dem_windows(self):
         """A basic comparison between the generated and benchmark DEM"""
@@ -145,6 +225,7 @@ class MultipleLidarGroundOnlyTest(unittest.TestCase):
     def test_result_dem_linux(self):
         """A basic comparison between the generated and benchmark DEM"""
 
+        # load in benchmark DEM
         file_path = self.cache_dir / self.instructions["data_paths"]["benchmark_dem"]
         with rioxarray.rioxarray.open_rasterio(file_path, masked=True) as benchmark_dem:
             benchmark_dem.load()
@@ -162,14 +243,15 @@ class MultipleLidarGroundOnlyTest(unittest.TestCase):
         threshold = 10e-6
         self.assertTrue(
             len(diff_array[diff_array != 0]) < len(diff_array) / 100,
-            f"{len(diff_array[diff_array != 0])} or more than 1% of DEM values differ "
-            f"on Linux test run: {diff_array[diff_array != 0]}",
+            f"{len(diff_array[diff_array != 0])} or more than 1% of DEM values differ"
+            f" on Linux test run: {diff_array[diff_array != 0]}",
         )
+        number_under_threshold = len(diff_array[numpy.abs(diff_array) > threshold])
         self.assertTrue(
-            len(diff_array[numpy.abs(diff_array) > threshold]) < len(diff_array) / 250,
-            "More than 0.4% of DEM values differ by more than {threshold} on Linux test"
-            f" run: {diff_array[numpy.abs(diff_array) > threshold]} or "
-            f"{len(diff_array[numpy.abs(diff_array) > threshold]) / len(diff_array.flatten()) * 100}%",
+            number_under_threshold < len(diff_array) / 250,
+            f"More than 0.4% of DEM values differ by more than {threshold} on Linux"
+            f" test run: {diff_array[numpy.abs(diff_array) > threshold]} or "
+            f"{number_under_threshold / len(diff_array.flatten()) * 100}%",
         )
 
         # explicitly free memory as xarray seems to be hanging onto memory
