+++ tests/test_dem_generation_westport_4/data/benchmark_dem.nc
